Close #36 - Implement Ajax Features
