+++ WebApplication/Site.Master.cs
@@ -10,8 +10,6 @@ namespace WebApplication {
 
 	public partial class SiteMaster : MasterPage {
 
-		public ConnectedUsersTrack ConnectedUsers => (ConnectedUsersTrack)Application.Get("LoggedUsers");
-
 		protected void Page_Load(object sender, EventArgs e) {
 
 			if(Convert.ToBoolean(Session["IsLogged"])) {
@@ -32,8 +30,6 @@ namespace WebApplication {
 
 			}
 
-			RefreshConnectedUsers();
-
 		}
 
 		protected void ButtonProfile_Click(object sender, EventArgs e) {
@@ -41,11 +37,8 @@ namespace WebApplication {
 		}
 
 		protected void ButtonSignOut_Click(object sender, EventArgs e) {
-			//Session["IsLogged"] = false;
-			//Session["Email"] = "";
-			//Session["Name"] = "";
-			//Session["LastName"] = "";
-			//Response.Redirect(AppConfig.WebSite.MainPage);
+			Session["IsLogged"] = false;
+			Response.Redirect(AppConfig.WebSite.MainPage);
 			Session.Abandon();
 		}
 
@@ -63,22 +56,6 @@ namespace WebApplication {
 					teacherMenu.Visible = true;
 					break;
 			}
-
-		}
-
-		protected void ConnectedUsersTimer_Tick(object sender, EventArgs e) {
-			RefreshConnectedUsers();
-		}
-
-		private void RefreshConnectedUsers() {
-
-			List<string> connectedStudents = ConnectedUsers.GetUsers("student").ToList();
-			ConnectedStudentCount.Text = connectedStudents.Count.ToString();
-			ConnectedStudents.DataSource = connectedStudents;
-
-			List<string> connectedTeachers = ConnectedUsers.GetUsers("teacher", "teacher_admin").ToList();
-			ConnectedTeacherCount.Text = connectedTeachers.Count.ToString();
-			ConnectedTeachers.DataSource = connectedTeachers;
 
 		}
