+++ WebApplication/UserPages/Teacher/Teacher.Master
@@ -5,37 +5,33 @@
 
 <asp:Content ID="BodyContent" ContentPlaceHolderID="MainContent" runat="server">
 
-	<div class="row">
+    <div class="row">
 
-		<!-- Nav tabs -->
-		<div class="col-3 nav-tabs-pills">
-			<div class="nav flex-column nav-pills nav-tabs" id="teacherTabs" role="tablist" aria-orientation="vertical">
-				<asp:LinkButton CssClass="nav-link" ID="HomeTab" PostBackUrl="/Teacher/Home" role="tab" aria-controls="Home" aria-selected="false" runat="server" CausesValidation="False">Home</asp:LinkButton>
-				<asp:LinkButton CssClass="nav-link" ID="TasksTab" PostBackUrl="/Teacher/Tasks" role="tab" aria-controls="Tasks" aria-selected="false" runat="server" CausesValidation="False">Tasks</asp:LinkButton>
-				<asp:Panel ID="TeacherAdminPanel" runat="server">
-					<asp:LinkButton CssClass="nav-link" ID="ImportTasksXmlDocumentTab" PostBackUrl="/Teacher/Tasks/ImportTasksXmlDocument" role="tab" aria-controls="ImportTasksXmlDocument" aria-selected="false" runat="server" CausesValidation="False">Import Tasks (XmlDocument)</asp:LinkButton>
-					<asp:LinkButton CssClass="nav-link" ID="ImportTasksDataSetTab" PostBackUrl="/Teacher/Tasks/ImportTasksDataSet" role="tab" aria-controls="ImportTasksDataSet" aria-selected="false" runat="server" CausesValidation="False">Import Tasks (DataSet)</asp:LinkButton>
-					<asp:LinkButton CssClass="nav-link" ID="ExportTasksTab" PostBackUrl="/Teacher/Tasks/Export" role="tab" aria-controls="ExportTasks" aria-selected="false" runat="server" CausesValidation="False">Export Tasks</asp:LinkButton>
-				</asp:Panel>
-			</div>
-		</div>
+        <!-- Nav tabs -->
+        <div class="col-3 nav-tabs-pills">
+            <div class="nav flex-column nav-pills nav-tabs" id="teacherTabs" role="tablist" aria-orientation="vertical">
+                <asp:LinkButton CssClass="nav-link" ID="HomeTab" PostBackUrl="/Teacher/Home" role="tab" aria-controls="Home" aria-selected="false" runat="server" CausesValidation="False">Home</asp:LinkButton>
+                <asp:LinkButton CssClass="nav-link" ID="TasksTab" PostBackUrl="/Teacher/Tasks" role="tab" aria-controls="Tasks" aria-selected="false" runat="server" CausesValidation="False">Tasks</asp:LinkButton>
+                <asp:Panel ID="TeacherAdminPanel" runat="server">
+                    <asp:LinkButton CssClass="nav-link" ID="ImportTasksXmlDocumentTab" PostBackUrl="/Teacher/Tasks/ImportTasksXmlDocument" role="tab" aria-controls="ImportTasksXmlDocument" aria-selected="false" runat="server" CausesValidation="False">Import Tasks (XmlDocument)</asp:LinkButton>
+                    <asp:LinkButton CssClass="nav-link" ID="ImportTasksDataSetTab" PostBackUrl="/Teacher/Tasks/ImportTasksDataSet" role="tab" aria-controls="ImportTasksDataSet" aria-selected="false" runat="server" CausesValidation="False">Import Tasks (DataSet)</asp:LinkButton>
+                    <asp:LinkButton CssClass="nav-link" ID="ExportTasksTab" PostBackUrl="/Teacher/Tasks/Export" role="tab" aria-controls="ExportTasks" aria-selected="false" runat="server" CausesValidation="False">Export Tasks</asp:LinkButton>
+                </asp:Panel>
+            </div>
+        </div>
 
-		<!-- Tab panes -->
-		<div class="col-9">
-			<div class="tab-content" id="teacherTabContents" role="tablist">
+        <!-- Tab panes -->
+        <div class="col-9">
+            <div class="tab-content" id="teacherTabContents" role="tablist">
 
-				<asp:UpdatePanel ID="TeacherTasksUpdatePanel" runat="server">
-					<ContentTemplate>
-						<asp:UpdateProgress runat="server">
-							<ProgressTemplate>
-							</ProgressTemplate>
-						</asp:UpdateProgress>
-						<asp:ContentPlaceHolder ID="TeacherContent" runat="server" />
-					</ContentTemplate>
-				</asp:UpdatePanel>
+                <asp:UpdatePanel ID="TeacherTasksUpdatePanel" runat="server">
+                    <ContentTemplate>
+                        <asp:ContentPlaceHolder ID="TeacherContent" runat="server" />
+                    </ContentTemplate>
+                </asp:UpdatePanel>
 
-			</div>
-		</div>
-	</div>
+            </div>
+        </div>
+    </div>
 
 </asp:Content>
